better main page styles add fonts add more pages, but disable them for now
