+++ package.json
@@ -67,6 +67,7 @@
 		"html-webpack-plugin": "^3.2.0",
 		"mini-css-extract-plugin": "^0.8.0",
 		"optimize-css-assets-webpack-plugin": "^5.0.3",
+		"prop-types": "^15.7.2",
 		"sass": "^1.23.0",
 		"sass-loader": "^8.0.0",
 		"style-loader": "^1.0.0",
