+++ src/components/SkillsPage.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function SkillsPage() {
+	return (
+		<div className="page" id="page-skills">
+			<div>Skills</div>
+		</div>
+	);
+}
