+++ src/components/WorkPage.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function WorkPage() {
+	return (
+		<div className="page" id="page-work">
+			<div>Work</div>
+		</div>
+	);
+}
