+++ webpack/config.js
@@ -40,7 +40,7 @@ module.exports = {
 				]
 			},
 			{
-				test: /\.(ttf|eot|woff|woff2)$/,
+				test: /\.(ttf|eot|woff|woff2|svg)$/,
 				use: {
 					loader: 'file-loader',
 					options: {
